requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
opencv-python-headless==4.8.1.78
pillow==10.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
pytest==7.4.3
